rb_new: use upstream react-leaflet
See https://github.com/PaulLeCam/react-leaflet/pull/537 for context.
