+++ indico/modules/rb_new/client/js/common/map/RoomBookingMap.jsx
@@ -19,76 +19,13 @@ import _ from 'lodash';
 import PropTypes from 'prop-types';
 import React from 'react';
 import {connect} from 'react-redux';
-import Leaflet from 'leaflet';
-import {Map, TileLayer, Marker, Tooltip} from 'react-leaflet';
-import MarkerClusterGroup from 'react-leaflet-markercluster';
+import {Map, TileLayer} from 'react-leaflet';
 import {Overridable} from 'indico/react/util';
 import {selectors as configSelectors} from '../config';
-import * as mapSelectors from './selectors';
+import MapMarkers from './MapMarkers';
 
 import 'leaflet/dist/leaflet.css';
 import './RoomBookingMap.module.scss';
-
-
-function groupIconCreateFunction(cluster) {
-    const highlight = cluster.getAllChildMarkers().some(m => m.options.highlight);
-    return Leaflet.divIcon({
-        html: `<span>${cluster.getChildCount()}</span>`,
-        className: `marker-cluster marker-cluster-small rb-map-cluster ${highlight ? 'highlight' : ''}`,
-        iconSize: Leaflet.point(40, 40, true),
-    });
-}
-
-
-function _MapMarkers({rooms, clusterProps, hoveredRoomId}) {
-    const icon = Leaflet.divIcon({className: 'rb-map-marker', iconSize: [20, 20]});
-    const hoveredIcon = Leaflet.divIcon({className: 'rb-map-marker highlight', iconSize: [20, 20]});
-    const highlight = rooms.some(r => r.id === hoveredRoomId);
-
-    return (
-        !!rooms.length && (
-            <MarkerClusterGroup showCoverageOnHover={false}
-                                iconCreateFunction={groupIconCreateFunction}
-                                // key is used here as a way to force a re-rendering
-                                // of the MarkerClusterGroup
-                                key={highlight}
-                                {...clusterProps}>
-                {rooms.filter(({lat, lng}) => !!(lat && lng)).map(({id, name, lat, lng}) => (
-                    // we have to make the key depend on the highlighted state, otherwise
-                    // the component won't properly refresh (for some strange reason)
-                    <Marker key={id}
-                            position={[lat, lng]}
-                            icon={id === hoveredRoomId ? hoveredIcon : icon}
-                            highlight={id === hoveredRoomId}>
-                        <Tooltip direction="top">
-                            <span>{name}</span>
-                        </Tooltip>
-                    </Marker>
-                ))}
-            </MarkerClusterGroup>
-        )
-    );
-}
-
-_MapMarkers.propTypes = {
-    rooms: PropTypes.array,
-    clusterProps: PropTypes.object,
-    hoveredRoomId: PropTypes.number,
-    /** 'actions' may be used by plugins */
-    actions: PropTypes.objectOf(PropTypes.func).isRequired
-};
-
-_MapMarkers.defaultProps = {
-    rooms: [],
-    clusterProps: {},
-    hoveredRoomId: null
-};
-
-const MapMarkers = connect(
-    state => ({
-        hoveredRoomId: mapSelectors.getHoveredRoom(state)
-    })
-)(_MapMarkers);
 
 
 class RoomBookingMap extends React.Component {
