+++ package.json
@@ -51,7 +51,7 @@
     "react-dom": "^16.6.3",
     "react-final-form": "^3.6.7",
     "react-jsx-i18n": ">=0.1.2",
-    "react-leaflet": "github:pferreir/react-leaflet#indico",
+    "react-leaflet": "^2.1.4",
     "react-leaflet-markercluster": "^2.0.0-rc3",
     "react-modal": "^3.6.1",
     "react-redux": "^5.1.1",
